Added newline support

+++ src/drivers/console.c
@@ -59,7 +59,7 @@ static int kfmt(char *buf, const char *fmt, va_list args) {
 
       size_t len = strlen(s);
 
-      for (int i = 0; i < len; ++i) {
+      for (size_t i = 0; i < len; ++i) {
         *str++ = *s++;
       }
       break;
